+++ app/gui2/src/components/shared/AgGridTableView.vue
@@ -8,29 +8,25 @@ import {
   tsvTableToEnsoExpression,
   writeClipboard,
 } from '@/components/GraphEditor/clipboard'
+import { TextFormatOptions } from '@/components/visualizations/TableVisualization.vue'
 import { useAutoBlur } from '@/util/autoBlur'
-import '@ag-grid-community/styles/ag-grid.css'
-import '@ag-grid-community/styles/ag-theme-alpine.css'
 import type {
   CellEditingStartedEvent,
   CellEditingStoppedEvent,
+  ColDef,
+  ColGroupDef,
   ColumnResizedEvent,
   FirstDataRenderedEvent,
+  GetRowIdFunc,
+  GridApi,
   GridReadyEvent,
   RowDataUpdatedEvent,
   RowEditingStartedEvent,
   RowEditingStoppedEvent,
-  SortChangedEvent,
-} from 'ag-grid-community'
-import type {
-  ColDef,
-  ColGroupDef,
-  GetRowIdFunc,
-  GridApi,
   RowHeightParams,
+  SortChangedEvent,
 } from 'ag-grid-enterprise'
 import { type ComponentInstance, reactive, ref, shallowRef, watch } from 'vue'
-import { TextFormatOptions } from '../visualizations/TableVisualization.vue'
 
 const DEFAULT_ROW_HEIGHT = 22
 
@@ -59,7 +55,7 @@ const gridApi = shallowRef<GridApi<TData>>()
 const popupParent = document.body
 useAutoBlur(() => grid.value?.$el)
 
-function onGridReady(event: GridReadyEvent) {
+function onGridReady(event: GridReadyEvent<TData>) {
   gridApi.value = event.api
 }
 
@@ -94,15 +90,15 @@ watch(
 )
 
 function updateColumnWidths(event: FirstDataRenderedEvent | RowDataUpdatedEvent) {
-  if (event.columnApi == null) {
-    console.warn('AG Grid column API does not exist.')
+  if (event.api == null) {
+    console.warn('AG Grid API does not exist.')
     return
   }
-  const cols = event.columnApi.getAllGridColumns().filter((c) => {
+  const cols = event.api.getAllGridColumns().filter((c) => {
     const id = c.getColId()
     return id && !widths.has(id)
   })
-  event.columnApi.autoSizeColumns(cols)
+  event.api.autoSizeColumns(cols)
 }
 
 function lockColumnSize(e: ColumnResizedEvent) {
@@ -206,6 +202,8 @@ const { AgGridVue } = await import('ag-grid-vue3')
   />
 </template>
 
+<style src="@ag-grid-community/styles/ag-grid.css" />
+<style src="@ag-grid-community/styles/ag-theme-alpine.css" />
 <style scoped>
 .ag-theme-alpine {
   --ag-grid-size: 3px;